requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
djangorestframework==3.16.1
idna==3.11
psycopg2-binary==2.9.11
requests==2.32.5
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.6.0
gunicorn==21.2.0
python-dotenv==1.0.0
dj-database-url==2.1.0
